Refactor functions.php and users.php
This commit refactors the functions.php file by adding a new function called current_user() that retrieves the current user's information from the database. It also adds a new function called find_by_groupLevel() that checks if a user belongs to a specific user level. Additionally, the page_require_level() function is modified to redirect users to the home page if they don't meet the required user level.

In the users.php file, the code is modified to remove the unnecessary fetching of users for the current page.